+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-* Removed `info` request after a file upload
+* Skipped `info` request after a file upload if the secret kay is missing
 
 ## 4.3.2 — 2023-03-28